AI-ENHANCED Hospital Resource Management System/ai-service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
numpy==1.25.2
pandas==2.1.0
scikit-learn==1.3.0
tensorflow==2.13.0
prophet==1.1.4
joblib==1.3.2
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
statsmodels==0.14.0